+++ boltdb/state.go
@@ -12,6 +12,7 @@ type stateRepository struct {
 	client     *bolt.DB
 	bucketName []byte
 	bucket     *bolt.Bucket
+	key        []byte
 }
 
 // NewStateRepository returns an implementation of the interface state.Repository
@@ -23,12 +24,13 @@ func NewStateRepository(c *bolt.DB) (state.Repository, error) {
 	return &stateRepository{
 		client:     c,
 		bucketName: bn,
+		key:        []byte("state"),
 	}, nil
 }
 
-func (r *stateRepository) Find(ctx context.Context, vid string) (*state.State, error) {
+func (r *stateRepository) Find(ctx context.Context) (*state.State, error) {
 	var s state.State
-	b := r.bucket.Get([]byte(vid))
+	b := r.bucket.Get(r.key)
 	if b == nil {
 		return &state.State{}, nil
 	}
@@ -39,10 +41,19 @@ func (r *stateRepository) Find(ctx context.Context, vid string) (*state.State, e
 	return &s, nil
 }
 
-func (r *stateRepository) Update(ctx context.Context, vid string, s *state.State) error {
+func (r *stateRepository) Update(ctx context.Context, s *state.State) error {
 	b, err := json.Marshal(s)
 	if err != nil {
 		return err
 	}
-	return r.bucket.Put([]byte(vid), b)
+	return r.bucket.Put(r.key, b)
+}
+
+func (r *stateRepository) DeleteAll(ctx context.Context) error {
+	bk, err := recreateBucket(r.bucket, r.bucketName)
+	if err != nil {
+		return err
+	}
+	r.bucket = bk
+	return nil
 }