+++ file/repository.go
@@ -9,4 +9,5 @@ type Repository interface {
 	CreateOrReplace(ctx context.Context, f *File) error
 	FindBySignature(ctx context.Context, sig string) (*File, error)
 	DeleteBySignature(ctx context.Context, sig string) error
+	DeleteAll(ctx context.Context) error
 }