+++ mock/file_repository.go
@@ -49,6 +49,20 @@ func (mr *FileRepositoryMockRecorder) CreateOrReplace(arg0, arg1 interface{}) *g
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateOrReplace", reflect.TypeOf((*FileRepository)(nil).CreateOrReplace), arg0, arg1)
 }
 
+// DeleteAll mocks base method.
+func (m *FileRepository) DeleteAll(arg0 context.Context) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "DeleteAll", arg0)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// DeleteAll indicates an expected call of DeleteAll.
+func (mr *FileRepositoryMockRecorder) DeleteAll(arg0 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeleteAll", reflect.TypeOf((*FileRepository)(nil).DeleteAll), arg0)
+}
+
 // DeleteBySignature mocks base method.
 func (m *FileRepository) DeleteBySignature(arg0 context.Context, arg1 string) error {
 	m.ctrl.T.Helper()